requirements.txt
python-telegram-bot==20.7
flask==2.3.3
psycopg2-binary==2.9.7
requests==2.31.0
openai==0.28.0
apscheduler==3.10.1
python-dotenv==1.0.0
gunicorn==21.2.0